Rename to asRetrievedVar() and isRetrieved (#54)
## What is the goal of this PR?
Clearer naming when indicating a variable is to be returned as part of the `get` clause. This is because in many places `asUserDefined` is not actually a user defined variable (especially in reasoner and QP) but instead used to ensure the variable is part of the `get` clause. This has lead to confusion about whether it's really about being able to check whether a variable was originally created by the user, or whether it's means to be returned in the `get`. The two main known use cases are:
1. Users creating variables with `var()`, then setting `asUserDefined()` to avoid creating a name for a variable they want to see in the `get()`
2. Internally, explicitly wanting to have the variable be returned.
Both boil down to creating anonymous variables that should be returned.

## What are the changes implemented in this PR?
Rename `asUserDefined()` to `asReturnedVar()` and `isUserDefinedName()` to `isReturned()`. This closes #53.