Restrict the token patterns for type labels and ID values (#60)
## What is the goal of this PR?

In the previous definition of the Graql grammar, we had a "leak" in the token pattern definition of type labels and ID values. Both of them were defined by the same token `identifier`, which allowed both `ID_` patterns and `STRING_` patterns. `STRING_` pattern literally means anything. We've refined the grammar to restrict the definition of type labels and ID values, by directly assigning `type_label` and `ID_` tokens wherever they are expected, and removed the `identifier` token which allowed `STRING_` tokens to be provided. This change results in a more strict and consistent grammar, and therefore less error-prone.

## What are the changes implemented in this PR?

- Refactored all of the token patterns for type labels (to have names beginning with `type_`
- Made `ID_` be independent of the definition of type labels
- Allowed `ID_` pattern to be a subset of patterns allowed for `type_name`
- Refactored the `Parser` to adapt to the new grammar
- Changed all query `.toString()` output to no longer put type labels and ID values in quotes (`"` or `'`)
- Made sure the token names follow the convention we've set (all tokens that describe a pattern beyond a literal value has an underscore `_` suffix)
- Grouped the `TYPE_CHAR` `fragments` with other fragment definitions