Renamed 'ValueType' to 'ValueClass' and 'date' to 'datetime' (#126)
## What is the goal of this PR?

With the new hypergraph `AttributeType` implementation, we no longer need the `Enum` `ValueType` in the server. We just pass a Java class such as `Boolean.class`, `Long.class`, etc, whenever needed - which is very rare. So the term `ValueType` (that previously refers to the enum containing a class) is now being changed again, to `ValueClass` (referring to the explicit Java class). Additionally, I realise this would be even better for understandability, as the suffix `Type` implies that it is a "Grakn Type", which therefore should be inheritable and have instances - which is not true for `ValueClass`.

Additionally, we're also renaming `date` to `datetime` to accurately describe the actual class it represents, and also allow us to introduce a `date` value class to represent explicitly only "dates".

## What are the changes implemented in this PR?

- Renamed the `ValueType` class to `ValueClass`
- Renamed method names
- Renamed method arguments
- Renamed variable names
- Updated comments