Add labels and scoped roles to the grammar and parsers (#140)
## What is the goal of this PR?

To being changes decided in #107 , we modify the grammar to have `type_scoped` as well as the default `type` that is unscoped. To make this make sense at the object level, we create a new object called `Label` that we will soon remove from `core` and `client-java` that can be scoped and unscoped types - Labels are now used throughout Graql to represent types. Where possible, we provide helper methods that consume `String`, creating the required `Label` objects internally and passing the Labels as required.

## What are the changes implemented in this PR?

* Create `Label`, a class holding the combination of scope + name. `scope` may be null
* The parser creates `Label` classes instead of strings from `label_ref`, `label_scoped`, and `label` rules
* `Compute` targets such as `of` and `in` require `Label` objects under the hood, though we provide convenience helpers that take strings and wrap them as `Label` where possible