Refactor Graql Predicates and Concept Equality (#168)
## What is the goal of this PR?

We would like to streamline the attribute value constraints to be more consistent. Right now, `>`, `<`, `>=`, and`<=`, all operate on _"attribute values"_. However, `!=` operates on _"concept equality"_. And thus to evaluate _"attribute value equality"_ and _"inequality"_, we must use `==` and `!==`. This is very inelegant and thus could be error-prone as they are very close `=` and `!=`. This behaviour is also inelegant for another reason: that it introduces a second way to do _"negation"_, which is not through our `not` logical operator. This issue has also be in raised through issue #167. 

Our conclusion is to remove _"concept inequality"_ constraint (`!=`), and introduce new constraint operator: `is`. `is` would allow us to assert that 2 concepts are equal: `$x is $y`, which we can then negate by doing `not { $x is $y };`.

Given the change above, we can now use `=` and `!=` that have been freed up, as _"attribute equality"_ and _"inequality"_.

As we go through this refactor, we also found that the `ValueOperation`, `ValueAssignment`, and `ValueComparison` data structures can be simplified into one `ValueConstraint`.

## What are the changes implemented in this PR?

1. Replace _"concept inequality"_ (`!=`) with _"concept equality"_ (`is`) in the Graql grammar
2. Simplify `value` definition in the Graql grammar
3. Replaced `EQV` (`==`) and `NEQV` (`!=`) with `EQ` (`=`) and `NEQ` (`!=`)
4. Replaced all usage of `ValueAssignment` with `ValueComparison + EQ`, inlined `ValueOperation` with `ValueComparison`, renamed `ValueComparison` to `ValueConstraint`
5. Update `Parser` to parse for `is` instead of `!=`
6. Update `Parser` to parse value assignment as value comparison
7. Update `Graql` builders to replace `.val(...)` with `.eq(...)`