Scope unscoped role types in Plays TypeConstraints (#190)
## What is the goal of this PR?

When a role type is overridden within the Plays constraint, the overridden type is specified with no scope- we can infer the scope based on it being a supertype of the supplied role type, but at parse time this is not possible to infer without building the entire graph tree. In order to avert this, we supplied the root Relation type as a scope for the relation so it can be identified as a role type ahead of time, which leads to it being passed over for definition.

## What are the changes implemented in this PR?

We added the root RelationType as the scope for overridden RoleTypes in the Plays TypeConstraint.
We changed the way that the Plays TypeConstraint reports its RoleTypes to match this inferred type.