Mark Isas dervied from has constraints as derived (#195)
## What is the goal of this PR?

In order to handle deletion properly, we need to be able to distinguish Isa constraints placed implicitly as part of has. For example, `$x has name $y` implies `$y isa name`, but if we are deleting `$x has name $y` we should not also delete $y's isa constraint. To this end, we added a flag on the Isa constraint indicating if it was derived.

## What are the changes implemented in this PR?

Added derived boolean to Isa constraint, set only when creating an isa implicitly via the Has constraints.