Strip leading and trailing whitespace prior to parsing (#8)
## What is the goal of this PR?

Previously, if a TypeQL string contained a syntax error, and also had a trailing newline, an unhelpful error was thrown: `Index 1 out of bounds for length 1`. This was due to a bug in the error handler.

Now, a more helpful error is reported, for instance:
```
There is a syntax error at line 1:
match $x isa thing
                  ^
no viable alternative at input 'match $x isa thing'
```

## What are the changes implemented in this PR?

Suppose we are parsing the following query:
```typeql
match $x isa thing

```
It includes a trailing newline.

Previously, we were attempting to parse the raw unmodified query string. The parser scans to the end of the query string and reports that there is a syntax error on line 2:
```
[TQL03] TypeQL Error: There is a syntax error at line 2:

^
no viable alternative at input 'match $x isa thing\n'
```
Firstly, this is still not a great error - it shows us that the "syntax error" is on a completely empty line.

Secondly, our error handler itself throws an error while trying to handle this error. In the error handler we convert the query string to a `List<String` using `String.split("\n")` with no `limit` parameter, which means that if the final character is a newline, we end up with less lines than we started with; in the example above, we have a 2-line query but only 1 entry in our String list.

We then print out a line from our string list that matches the line number where the error was reported, and boom, we get our `ArrayIndexOutOfBoundsException`.

We _could_ fix it by using `String.split("\n", -1)` which preserves all lines, even if they are blank; but, while that intuitively feels like the more "correct" fix, in practice, the resulting error is much nicer if we actually strip out the trailing newline. So we now use `trim` to strip out all leading and trailing whitespace, which should be intuitive enough for users.

(One thing to consider: should we strip only **trailing** whitespace?)