Remove obsolete tests (#226)
## What is the goal of this PR?

We remove a test that used to test functionality that is no longer supported, namely, `match` queries with no named variables.

## What are the changes implemented in this PR?

`testRoleTypeScopedSpecifically`: The test used to centre around the `match marriage relates spouse;` query at its conception. This is no longer supported, and the updated version is exactly equivalent to the preceding test.
`testQueryParserWithoutGraph`: This test was present in the initial commit, and it looks like it was testing that we can call a static method without referencing its class. Maybe it had a different use at some point prior, but that is lost to history.