Introduce unformatted toString (#239)
## What is the goal of this PR?

We introduce an unformatted `toString` method which will not introduce
any indentation or newlines.

This fixes a bug where a user's strings with newlines in them receive
indentation in the middle of the attribute. Now, to avoid this visual or
downstream effect the API exists to disable newlines and indentation.

## What are the changes implemented in this PR?

* introduce `toString(boolean pretty)` which allows enabling or
disabling pretty-printing. This API is available on `Defineable`,
`TypeQLQuery`, and `Pattern`, but not `Constraint` since these are very
short and don't do any pretty printing themselves.
* the default `toString()` on all objects does do pretty printing when
the option is available