Bump dependencies and python version for GitHub deployment (#248)
## What is the goal of this PR?

We've fixed the Python version learned from the successful GitHub deployment of typedb-common a definite working version of Python for this job.

## What are the changes implemented in this PR?

We've bumped typedb-common and typedb-behaviour. We've also changed the used Python version to 3.7.9 from 3.7.12. This fixes a later bug where we attempt to link a library to a not-installed version of Python.