Add Rust IDE sync tool (#247)
## What is the goal of this PR?

We added the Rust IDE sync tool to enable Rust development in IDEs.

## What are the changes implemented in this PR?

- Bump dependencies
- Add `dependencies/ide/rust/sync.sh`
- Add Cargo files and build artifacts to `.gitignore`

### Usage

Run `./dependencies/ide/rust/sync.sh`, immediately after cloning/pulling
this repo, or after loading it in an IDE.

This will generate `Cargo.toml` files in the project, allowing the Rust
targets to be built, run and debugged using Cargo. The Vaticle fork of
the IntelliJ Rust plugin is required: this modified plugin is able to
run the `cargo` binary that is loaded into the Bazel sandbox on building
the Bazel project.