Add transitive Rust deps that aren't included by default on M1 Macs (#250)
## What is the goal of this PR?

We added `core-foundation-sys` and `libc` explicitly as Rust
dependencies. They are conditional transitive dependencies of
`antlr-rust` and `chrono` respectively. Due to
https://github.com/vaticle/dependencies/issues/385 they are not included
by default.

## What are the changes implemented in this PR?

`core-foundation-sys` and `libc` are conditional dependencies of the
`antlr-rust` and `chrono` crates. The condition is the Cargo condition
`cfg('unix')`, which is supposed to represent any Unix system. Due to
https://github.com/vaticle/dependencies/issues/385, they are not
correctly pulled in when the machine has `aarch64` architecture. For
now, we work around the issue by declaring them as explicit
dependencies.