Rusty error messages (#258)
## What is the goal of this PR?

We replace the `Message` error struct with the `TypeQLError` enum, which greatly simplifies error handling on the user side.

## What are the changes implemented in this PR?

The `error_messages!` macro used to generate message templates has been replaced with one that generates an error enum instead. Returning an enum as the error is the established Rust approach as it greatly simplifies error handling. We also export the error generator macro so that it can be reused in e.g. typedb-client-rust.

Another minor change is adding the `group_imports = "StdExternalCrate"` to `rustfmt.toml`, a change backported from client-rust.