Sanitise rust macros to avoid forcing user's environment (#299)
## What is the goal of this PR?

The problem is described in #298: we have macros that need access to other
items in our crate. Paths to these items were not absolute, and users
had to import that particular items and were able to use their own
objects with the same names and use it in the macros. Now all paths are
absolute, based on the `$crate` metavariable.

## What are the changes implemented in this PR?

- We updated all macros that need access to other items in the crate.
- We removed redundant imports. 

closes #298