Allow variables to have a leading digit (#317)
## Usage and product changes

We modify the behaviour of #310 which unified variables and labels to
have the same valid identifier syntax, which eliminated the capability
of variables to have a leading number. For example, the variable `$0`
was banned.

This PR reverts this specific behaviour, and enables usage of variables
with leading digits:
```
match
$1_a isa entity;
get;
```
is made valid again.

Testing specified in
https://github.com/vaticle/typedb-behaviour/pull/281

## Implementation

We split the identifier regexes and grammar rules that validated labels
and variables before, to be specific to `Variable` and `Label`
identifiers.