Add query end marker (#405)
## Product change and motivation

TypeQL query pipelines cannot be distinguished since query stages can be
arbitrarily chained and concatenated. To resolve this, we introduce the
`end;` marker. For example:
```
match ...
insert ...
end;
match ...
insert...
```

Is now correctly interpretable as two separate match-insert queries!
Compared to before:
```
match ...
insert ...
match ...
insert ...
```

Where it was not clear if this is one match-insert-match-insert
pipeline, two match-insert pipelines, or 4 separate queries, or anything
in between!

For simplicity, any query can be terminated with an 'end;` marker,
though it is redundant in for schema queries and 'fetch' queries:
```
define ...;
end; # redundant!

match ...
insert ...
end;  # not redundant!

match ...
insert ...
fetch { ... };
end;  # redundant!
```

## Implementation

We introduce the 'end' marker that allows terminating any query, even if
it is a redundant marker.

The parsed `Query` type now also contains a boolean indicating whether
the query was explicitly ended with an 'end;' marker.