Remove Java and update README (#409)
## Product change and motivation
Remove deprecated Java code and grammar and update READMEs to align with
the recent changes in TypeDB 3.x.

## Motivation
The outdated codebase confused contributors and other readers.

## Implementation
We remove the whole Java codebase from the repository. It is still
accessible through tags and releases in case we want to reuse some of
the code in future implementations of Java query builders and ORMs. As
we still aim to support Java, no architectural changes (Bazel
configuration and tooling) have been touched, so it will be simpler to
reintroduce these elements in the future.

We update READMEs to actualize the language and references.